requirements.txt
discord.py
pytest
requests
